pyproject.toml
[project]
name = "pasaigen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "instructor[litellm]>=1.10.0",
    "litellm>=1.75.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
